Added script/conditional chapter.
Also moved the script tags into a shortcode at the bottom of the
document, so as not to be a total hypocrite about where I tell
people to load things.

Removed the "serve" action from grunt.js, since it's not really
necessary for the way I'm loading files.

Added Google Code syntax highlighting (remote for now).